Move encode decode tests to a single files
