+++ packages/lodestar/test/unit/network/reqresp/utils.ts
@@ -1,5 +1,8 @@
-import {Root, Status} from "@chainsafe/lodestar-types";
+import {expect} from "chai";
+import {Root, SignedBeaconBlock, Status} from "@chainsafe/lodestar-types";
+import {LodestarError} from "@chainsafe/lodestar-utils";
 import {List} from "@chainsafe/ssz";
+import {generateEmptySignedBlock} from "../../../utils/block";
 
 export function createStatus(): Status {
   return {
@@ -27,11 +30,22 @@ export function isEqualSszType<T>(type: {equals: (a: any, b: any) => boolean}, a
 }
 
 /**
- * Helper for it-pipe when first argument is an array
+ * Helper for it-pipe when first argument is an array.
  * it-pipe does not convert the chunks array to a generator and BufferedSource breaks
  */
 export async function* arrToSource<T>(arr: T[]): AsyncGenerator<T> {
   for (const item of arr) {
     yield item;
   }
+}
+
+export function expectLodestarError<T extends {code: string}>(err1: LodestarError<T>, err2: LodestarError<T>): void {
+  if (!(err1 instanceof LodestarError)) throw Error(`err1 not instanceof LodestarError: ${(err1 as Error).stack}`);
+  if (!(err2 instanceof LodestarError)) throw Error(`err2 not instanceof LodestarError: ${(err2 as Error).stack}`);
+
+  expect(err1.getMetadata()).to.deep.equal(err2.getMetadata(), "Wrong LodestarError metadata");
+}
+
+export function generateEmptySignedBlocks(n = 3): SignedBeaconBlock[] {
+  return Array.from({length: n}).map(() => generateEmptySignedBlock());
 }
